Fix a regression that caused DHCP to fail

+++ entrypoint.d/chrony.sh
@@ -3,8 +3,9 @@
 # Chrony configuration helpers for Domain-In-A-Box.
 
 dib_configure_chrony() {
-    echo "Writing /etc/chrony/chrony.conf..."
-    tee /etc/chrony/chrony.conf >/dev/null <<EOF
+    if [ ! -f /etc/chrony/chrony.conf ]; then
+        echo "Writing /etc/chrony/chrony.conf..."
+        tee /etc/chrony/chrony.conf >/dev/null <<EOF
 pool pool.ntp.org iburst
 driftfile /var/lib/chrony/chrony.drift
 makestep 1.0 3
@@ -15,4 +16,7 @@ bindaddress ${IP}
 bindcmdaddress 127.0.0.1
 ntpsigndsocket /var/lib/samba/ntp_signd
 EOF
+    else
+        echo "Keeping existing /etc/chrony/chrony.conf"
+    fi
 }
